ocamlyices2: function 'main' in yices2.ml may not be called
The main function in yices2.ml registers the ocaml exception that is
raised from the C code. It is important that this function is called
before any C code is run.

The thing is that in some cases (I ran into), the 'main' function from
the ocamlyices2 library is not called. In this commit, I propose a way
for the user to call the function himself when the exception is 'not found'.
This issue doesn't appear in the example or in the tests.

This bug appeared in the touist project.