libyices: smt2_commands.h: avoid symbol conflicts by adding 'typedef'
Problem: smt2_commands.c and smt2_parser.c and smt2_term_stack.c
all include smt2_commands.h. But in this header, an enum is defined.
When these three files are compiled into three .o and then linked together,
the symbol 'smt2_errors' are conflicting.

Using 'typedef' is a straightforward fix. We could also fix that by
moving the definition of the enum in the smt2_commands.c
and only keep a 'extern enum' in the .h file.