libyices: csl_check_libs: use ./libtool with shared library enabled
Because the ./libtool is generated depending on --disable/enable-shared,
we must use a different ./libtool with shared library enabled in any case.
This is useful in the case of

    ./configure --with-pic --disable-shared --with-pic-gmp=$PWD/with-pic/libgmp.a

which means that we want to wheck that with-pic/libgmp.a is PIC (and thus
we want ./libtool to accept to build shared libraries, which is the way I use
for checking the PIC property).